12.1.Let's Make a Console Command! {php bin/console make:command}
We've created exactly one email... and done some pretty cool stuff with it. Let's introduce a second email... but with a twist: instead of sending this email when a user does something on the site - like register - we're going to send this email from a console command. And that... changes a few things.
Let's create the custom console command first. Here's my idea: one of the fields on User is called $subscribeToNewsletter. In our pretend app, if this field is set to true for an author - someone that writes content on our site - once a week, via a CRON job, we'll run a command that will email them an update on what they published during the last 7 days.

Making the Command

Let's bootstrap the command... the lazy way. Find your terminal and run:

php bin/console make:command
Call it app:author-weekly-report:send. Perfect! Back in the editor, head to the src/Command directory to find... our shiny new console command.