12.2.Let's Make a Console Command! {Making the Command}
Let's start customizing newly created file: we don't need any arguments or options... and I'll change the description:

  Send weekly reports to authors.

The first thing we need to do is find all users that have this  property set to true in the database. To keep our code squeaky clean, let's add a custom repository method for that in UserRepository. How about public function findAllSubscribedToNewsletter(). This will return an array. Inside, return $this->createQueryBuilder(), u as the alias, ->andWhere('u.subscribeToNewsletter = 1'), ->getQuery() and ->getResult(). Above the method, we can advertise that this specifically returns an array of User objects.