12.3.Let's Make a Console Command! {Autowiring Services into the Command}
Autowiring Services into the Command

Back in the command, let's autowire the repository by adding a constructor. This is one of the rare cases where we have a parent class... and the parent class has a constructor. I'll go to the Code -> Generate menu - or Command + N on a Mac - and select "Override methods" to override the constructor.

Notice that this added a  argument - that's an argument in the parent constructor - and it called the parent constructor. That's important: the parent class needs to set some stuff up.