12.4.Let's Make a Console Command! {Autowiring Services into the Command}
But, we don't need to pass the command name: Symfony already gets that from a static property on our class. Instead, make the first argument: UserRepository . Hit Alt + Enter and select "Initialize fields" to create that property and set it. Perfect. Next, in execute(), clear everything out except for the $io variable, which is a nice little object that helps us print things and interact with the user... in a pretty way.
Start with $authors = $this->userRepository->findAllSubscribedToNewsletter(). Well, this really returns all users... not just authors - but we'll filter them out in a minute. To be extra fancy, let's add a progress bar! Start one with ->progressStart(). Then, foreach over  as , and advance the progress inside.
Oh, and of course, for progressStart(), I need to tell it how many data points we're going to advance. Use count($authors). Leave the inside of the foreach empty for now, and after, say $io->progressFinish(). Finally, for a big happy message, add $io->success()

  Weekly reports were sent to authors!
Brilliant! We're not doing anything yet... but let's try it! Copy the command name, find your terminal, and do it!

php bin/console app:author-weekly-report:send

Super fast