12.5.Let's Make a Console Command! {Counting Published Articles}
Counting Published Articles

Inside the foreach, the next step is to find all the articles this user published - if any - from the past week. Open up ArticleRepository... and add a new method for this - findAllPublishedLastWeekByAuthor() - with a single argument: the User object. This will return an array... of articles: let's advertise that above.
The query itself is pretty simple: return $this->createQueryBuilder() with ->andWhere('a.author = :author) to limit to only this author - we'll set the :author parameter in a second - then ->andWhere('a.publishedAt > :week_ago'). For the placeholders, call setParameter() to set author to the $author variable, and ->setParameter() again to set week_ago to a new \DateTime('-1 week'). Finish with the normal ->getQuery() and ->getResult().
Boom! Back in the command, autowire the repository via the second constructor argument: ArticleRepository $articleRepository. Hit Alt + Enter to initialize that field.
Down in execute, we can say $articles = $this->articleRepository->findAllPublishedLastWeekByAuthor() and pass that $author.
Phew! Because we're actually querying for all users, not everyone will be an author... and even less will have authored some articles in the past 7 days. Let's skip those to avoid sending empty emails: if count($articles) is zero, then continue.
By the way, in a real app, where you would have hundreds, thousands or even more users, querying for all that have subscribed is not going to work. Instead, I would make my query smarter by only returning users that are authors or even query for a limited number of authors, keep track of which you've sent to already, then run the command over and over again until everyone has gotten their update. These aren't even the only options. The point is: I'm being a little loose with how much data I'm querying for: be careful in a real app.
Ok, I think we're good! I mean, we're not actually emailing yet, but let's make sure it runs. Find your terminal and run the command again:

php bin/console app:author-weekly-report:send
All smooth. Next... let's actually send an email! And then, fix the duplication we're going to have between our two email templates.