13.1.Using a Base Email Template
We found all the authors that want to receive an update about the articles they wrote during the last 7 days. Now, let's send them that update as an email.
If you downloaded the course code, you should have a tutorial/ directory with an inky/ directory and a file inside called author-weekly-report.html.twig. Copy that and throw it into templates/email/.
Nice! This template is already written using the Inky markup: the markup that Inky will translate into HTML that will work in any email client. But mostly, other than a link to the homepage and the user's name, this is a boring, empty email: we still need to print the core content of the email.