13.2.Using a Base Email Template{Designing, Configuring & Sending that Email}
Designing, Configuring & Sending that Email

Let's open up welcome.html.twig, steal the apply line from here, and paste it on top of the new template. This will translate the markup to Inky and inline our CSS. At the bottom, add endapply... and I'll indent everything to satisfy my burning inner need for order in the universe!
To send this email, we know the drill! In the command, start with $email = (new TemplatedEmail()), ->from() and... ah: let's cheat a little.
Go back to src/Controller/SecurityController.php, find the register() method and copy its from() line: we'll probably always send from the same user. And yes, we'll learn how not to duplicate this later. I'll re-type the "S" on NamedAddress and hit tab to add the missing use statement on top.
Do we need to pass any variables to the template? Technically... no: the only variable we're using so far is the built-in email variable. But we will need the articles, so let's call ->context([]). Pass this an author variable... I'm not sure if we'll actually need that... and the $articles that this author recently wrote.
Done! Another beautiful Email object. We're a machine! How do we send it? Oh, we know that too: we need the mailer service. Add a third argument to the constructor: MailerInterface . I'll do our usual Alt+Enter trick and select "Initialize Fields" to create that property and set it.
Back down below, give a co-worker a serious "nod"... as if you're about to take on a task of great gravity... but instead, send an email: $this->mailer->send($email).
Love that. In our fixtures, thanks to some randomness we're using, about 75% of users will be subscribed to the newsletter. Before we run the command, let's make sure the data is fresh... with recent article created dates. Run:

php bin/console doctrine:fixtures:load
This should add enough users and articles that about 1-2 authors will be subscribed to the newsletter and have recent articles. Try that command:

php bin/console app:author-weekly-report:send
Ha! It didn't explode! It found 6 authors... or really, 6 users that are subscribed to the newsletter... but anywhere from 0 to 6 of these might actually have recent articles. Spin over to Mailtrap. If you don't see any emails - try reloading the fixtures again... just in case you got some bad random data, then re-run the command. Oh, and if you got an error when running the command about too many emails being sent, you've hit a limit on Mailtrap. The free plan only allows sending 2 emails each 10 seconds. In that case, ignore the error - because two emails did send - or reload your fixtures to hopefully send less emails.
We have exactly one email: phew! So... we rock! Or do we?
I see a few problems. First, the link to the homepage is broken: it links to localhost. Not localhost:8000 - or whatever our real domain is - just localhost. When you send emails from a console command... your paths break. More on that later.

Base Email Template

The second problem is more obvious... and it's my fault: this email is missing the cool header and footer we had in the other email! Why? Simple: in welcome.html.twig, we have a header with a logo on top and a footer at the bottom. In author-weekly-report.html.twig? I forgot to put that stuff!
Ok, I really did it on purpose: we probably do want a consistent layout for every email... but we definitely do not want to duplicate that layout in every email template.