13.3.Using a Base Email Template{Base Email Template}
Base Email Template

We definitely do not want to duplicate that layout in every email template.
We know the fix! We do it all the time in normal twig: create a base template, a base email template. In the templates/email directory, add a new file called, how about emailBase.html.twig.
And... I'll close a few files. In welcome.html.twig, copy that entire template and paste in emailBase. Then... select the middle of the template and delete! We basically want the header, the footer and, in the middle, a block for the content. Add {% block content %}{% endblock %}.
That block name could be anything. Now that we have this nifty template, back in welcome.html.twig, life gets simpler. On top, start with {% extends 'email/emailBase.html.twig' %}. Then, delete the apply and endapply, and replace it with {% block content %}... and {% endblock %}.
If you're wondering why we don't need the inky_to_html and inline_css filter stuff anymore, it's because the contents of this template will be put into a block that is inside of those same filters. The content will go through those filters... but we don't need to worry about adding them in every template.
Now we can delete most of the content: all we really need is the welcome row... and down below, we can get rid of the bottom and footer stuff. Celebrate your inner desire for order by un-indenting this.
Perfecto! Repeat this beautiful code in author-weekly-report.html.twig: {% extends 'email/emailBase.html.twig' %}, {% block content %} and all the way at the bottom, {% endblock %}. We can also remove the container element... and unindent.
That felt great! Let's see how it looks: run our weekly report:

php bin/console app:author-weekly-report:send
And... move back over! Woo! Now every email can easily share the same look.

Next, let's finish the email by making it dynamic. And, most importantly, let's figure out why our link paths are broken. You need to be extra careful when you send an email from the command line.