14.2.Router Request Context: Fix Paths in the CLI{Why is the Link Broken}
Why is the Link Broken

Now let's turn to the glaring, horrible bug in our email! Ah! As I mentioned a few minutes ago, if you hover over the link its, gasp, broken! For some reason, it points to localhost not our real domain... which is localhost:8000. Close, but not right.
Hmm. In the template... yea... that looks right: {{ url('app_homepage') }}. Ok, then why - when we click on the link - is it broken?
 41 lines  templates/email/author-weekly-report.html.twig
{% block content %}
    <row>
        <columns>
            <center>
                <button href="{{ url('app_homepage') }}">Check on the Space Bar</button>
            </center>
        </columns>
    </row>
{% endblock %}
We know that the url() function tells Symfony to generate an absolute URL. And... it is. I'll run "Inspect Element" on the broken link button. Check out the href: http://localhost not localhost:8000. The same thing would happen if you deployed this to production: it would always say localhost. The URL is absolute... it's just wrong!
Why? Think about it: in the registration email - where this did work - how did Symfony know what our domain was when it generated the link? Did we configure that somewhere? Nope! When you submit the registration form, Symfony simply looks at what the current domain is - localhost:8000 - and uses that for all absolute URLs.
But when you're in a console command, there is no request! Symfony has no idea if the code behind this site is deployed to localhost:8000, example.com, or lolcats.com. So, it just guesses localhost... which is totally wrong... but probably better than guessing lolcats.com?
If you're sending emails from the command line - or rendering templates for any reason that contain paths - you need to help Symfony: you need to tell it what domain to use.