14.3.Router Request Context: Fix Paths in the CLI{Setting router.request_context; Using Environment Variables... in Environment Variables}
Setting router.request_context

To fix this, start by looking inside our .env file. One of our keys here is called SITE_BASE_URL.
It is the URL to our app. But, but, but! This is not a standard Symfony environment variable and Symfony is not currently using this. Nope, this is an environment variable that we invented in our file uploads tutorial for a totally different purpose. You can see it used in config/services.yaml. It has nothing to do with how Symfony generates URLs.
Anyways, to fix the path problem, you need to set two special parameters. The first is router.request_context.scheme, which you'll set to https or http. The other is router.request_context.host which, for our local development, will be localhost:8000.
Now obviously, we don't want to hardcode these - at least not the second value: it will be different on production. Instead, we need to set these as new environment variables. And... hey! In .env, the SITE_BASE_URL is almost what we need... we just need it to be kind of split into two pieces. Hmm.
Check this out, create two new environment variables: SITE_BASE_SCHEME set to https and SITE_BASE_HOST set to localhost:8000. Back in services.yaml, use these values: %env(SITE_BASE_SCHEME)% and %env(SITE_BASE_HOST)%

Using Environment Variables... in Environment Variables

The problem is that we now have some duplication. Fortunately, one of the properties of environment variables is that... um... they can contain environment variables! For SITE_BASE_URL, set it to $SITE_BASE_SCHEME - yep, that's legal - :// and then $SITE_BASE_HOST
I love that trick. Anyways, now that we've set those two parameters, Symfony will use them to generate the URL instead of trying to guess it. Try the command one last time:

php bin/console app:author-weekly-report:send
And... check it out in Mailtrap! Yes! This time the link points to localhost:8000.
Next! Let's talk about attaching files to an email. Hmm, but to make it more interesting, let's first learn how to generate a styled PDF.