18.1.Organizing Emails Logic into a Service{Creating a Mailer Service}
We're sending two emails: one from a command and the other from src/Controller/SecurityController.php. The logic for creating and sending these emails is fairly simple. But even still, I prefer to put all my email logic into one or more services. The real reason for this is that I like to have all my emails in one spot. That helps me remember which emails we're sending and what they contain. After all, emails are a strange part of your site: you send a lot of them... but rarely or never see them! Like, how often do you do a "password reset" on your own site to check out what that content looks like? Keeping things in one spot... at least helps with this.

Creating a Mailer Service
So what we're going to do is, in the Service/ directory, create a new class called FileThatWillSendAllTheEmails... ah, or, maybe just Mailer... it's shorter.
The idea is that this class will have one method for each email that our app sends. Now, if your app sends a lot of emails, instead of having just one Mailer class, you could instead create a Mailer/ directory with a bunch of service classes inside - like one per email. In both cases, you're either organizing your email logic into a single service or multiple services in one directory.
Start by adding an __construct() method. The one service that we know we're going to need is MailerInterface $mailer... because we're going to send emails. I'll hit Alt + Enter and go to "Initialize fields" to create that property and set it.
Ok, let's start with the registration email inside of SecurityController. Ok... to send this email, the only info we need is the User object. Create a new public function sendWelcomeMessage() with a User $user argument.
Then, grab the logic from the controller... everything from $email = to the sending part... and paste that here. It looks like this class is missing a few use statements... so I'll re-type the "L" on TemplatedEmail and hit tab, then re-type the S on NamedAddress and hit tab once more... to add those use statements to the top of this file. Then change $mailer to $this->mailer.

 Tip

In Symfony 4.4 and higher, use new Address() - it works the same way as the old NamedAddress.
I love it! This will simplify life dramatically in SecurityController. Delete all the logic and then above... replace the MailerInterface argument with our shiny new Mailer class. Below, it's as simple as $mailer->sendWelcomeMessage($user). That looks really nice! Our controller is now more readable.