18.2.Organizing Emails Logic into a Service{Creating a Mailer Service}
Let's repeat the same thing for our weekly report email. In this case, the two things we need are the $author that we're going to send to - which is a User object - and the array of articles. Ok, over in our new Mailer class, add a public function sendAuthorWeeklyReportMessage() with a User object argument called $author and an array of Article objects.
Time to steal some code! Back in the command, copy everything related to sending the email... which in this case includes the entrypoint reset, Twig render, PDF code and the actual email logic. Paste that into Mailer.
This time, we need to inject a few more services - for entrypointLookup, twig and pdf. Let's add those on top: Environment $twig, Pdf $pdf and EntrypointLookupInterface $entrypointLookup. I'll do my Alt + Enter shortcut and go to "Initialize fields" to create those three properties and set them.
Back in the method... oh... that's it! We're already using the properties... and everything looks happy! Oh, and it's minor, but I'm going to move the "entrypoint reset" code below the render. This is subtle... but it makes sure that the Encore stuff is reset after we render our template. If some other part of our app calls this methods and then renders its own template, Encore will now be ready to do work correctly for them.
Anyways, let's use this in the command. Delete all of this logic and... in the constructor, change the  argument to Mailer $mailer. Now we get to delete stuff! Take off the $twig, $pdf and $entrypointLookup arguments, clear them from the constructor and remove their properties. If you really want to make things squeaky-clean, we now have a bunch of "unused" use statements that are totally useless.
Back down, call the method with $this->mailer->sendWeeklyReportMessage() passing $author and $articles.
Phew! This really simplifies the controller & command... and now I know exactly where to look for all email-related code. Let's... just make sure I didn't break anything. Run:

php bin/console app:author-weekly-report:send
No errors... and in Mailtrap... yep! 2 emails... with an attachment!
Next, sending emails is scary! So let's add some tests. We'll start by adding a unit test and later, an integration test, functional test... and a final exam that will be worth 50% of your grade for the semester. Ok... no final exam - but we will do that other stuff.