20.1.Integration Testing Emails
I also want to test the method that sends the weekly update email. But because the real complexity of this method is centered around generating the PDF, instead of a unit test, let's write an integration test.
In MailerTest, add a second method: testIntegrationSendAuthorWeeklyReportMessage().
Let's start the same way as the first method: copy all of its code except for the asserts, paste them down here and change the method to sendAuthorWeeklyReportMessage().
This needs a User object... but it also needs an array of articles. Let's create one: $article = new Article(). These articles are passed to the template where we print their title. So let's at least populate that property: $article->setTitle():

  Black Holes: Ultimate Party Pooper
Use this for the 2nd argument of sendAuthorWeeklyReportMessage(): an array with just this inside.

Unit Versus Integration Test

It's time to think strategically about our mocks. Right now, every dependency is mocked, which means it's a pure unit test. If we kept doing this, we could probably make sure that whatever render() returns is passed to the PDF function... and even assert that whatever that returns is passed to the attach() method. It's not bad, but because the logic in this method isn't terribly complex, its usefulness is limited.
What really scares me is the PDF generation: does my Twig template render correctly? Does the PDF generation process work... and do I really get back PDF content? To test this, instead of mocking $twig and $pdf, we could use the real objects. That would make this an integration test. These are often more useful than unit tests... but are also much slower to run, and it will mean that I really do need to have wkhtmltopdf installed on this machine, otherwise my tests will fail. Tradeoffs!
So here's the plan: use the real $twig and $pdf objects but keep mocking $symfonyMailer and $entrypointLookup... because I don't really want to send emails... and the $entrypointLookup doesn't matter unless I want to test that it does reset things correctly between rendering 2 PDFs.