20.2.Integration Testing Emails{Become an Integration Test! ;Fetching out Services; Caching Driver in the test Environment}
Become an Integration Test!

To make this test able to use real objects, we need to change extends from TestCase to KernelTestCase.
That class extends the normal TestCase but gives us the ability to boot Symfony's service container in the background. Specifically, it gives us the ability, down in the method, to say: self::bootKernel()
That will give us the ability to fetch real service objects and use them.

Fetching out Services

So we'll leave $symfonyMailer mocked, leave the $entrypointLookup mocked, but for the Pdf, get the real Pdf service. How? In the test environment, we can fetch things out of the container using the same type-hints as normal. So, $pdf = self::$container - bootKernel() set that property - ->get() passing this Pdf::class. Do the same for Twig: self::$container->get(Environment::class).
I love that! Again, the downside is that you really do need to have wkhtmltopdf installed correctly anywhere you run your tests. That's the cost of doing this.
Before we try it, at the bottom, we don't have any asserts yet. Let's add at least one: $this->assertCount() that 1 is the count of $email->getAttachments()
We could go further and look closer at the attachment... maybe make sure that it looks like it's in a PDF format... but this is a good start.
Now let's try this. Find your terminal and run our normal:

php bin/phpunit
It is slower this time... and then.. ah! What just happened? Two things. First, because this booted up a lot more code, we're seeing a ton of deprecation warnings. These are annoying... but we can ignore them.

Caching Driver in the test Environment

The second thing is that... the test failed! But... weird - not how I expected: something about APCu is not enabled. Huh? Why is it suddenly trying to use APCu?
The cause of this is specific to our app... but it's an interesting situation. Open up config/packages/cache.yaml.

 21 lines  config/packages/cache.yaml
framework:
    cache:
        app: '%cache_adapter%'
See this app key? This is where you can tell Symfony where it should store things that need to be added to cache at runtime - like the filesystem, redis or APCu. In an earlier tutorial, we set this to a parameter that we invented: %cache_adapter%.

This allows us to do something cool. Open config/services.yaml.

 53 lines  config/services.yaml
parameters:
    cache_adapter: cache.adapter.apcu
Here, we set cache_adapter to cache.adapter.apcu: we told Symfony to store cache in APCu. And... apparently, I don't have that extension installed on my local machine.
Ok... fine... but then... how the heck is the website working? Shouldn't we be getting this error everywhere? Yep... except that we override this value in services_dev.yaml - a file that is only loaded in the dev environment. Here we tell it to use cache.adapter.filesystem.

3 lines  config/services_dev.yaml
parameters:
    cache_adapter: 'cache.adapter.filesystem'
This is great! It means that we don't need any special extension for the cache system while developing... but on production, we use the superior APCu.
The problem now is that, when we run our tests, those are run in the test environment... and since the test environment doesn't load services_dev.yaml, it's using the default APCu adapter! By the way, there is a services_test.yaml file... but it has nothing in it. In fact, you can delete this: it's for a feature that's not needed anymore.